+++ gwcelery/conf/__init__.py
@@ -170,3 +170,7 @@ channel_names = {'H1': 'H1:GDS-CALIB_STRAIN',
                  'V1': 'V1:Hrec_hoft_16384Hz'}
 """Names of h(t) channels used in Parameter Estimation with LALInference (see
 :mod:`gwcelery.tasks.lalinference`)"""
+
+pe_threshold = 1.0 / (14 * 86400)
+"""FAR threshold in Hz for Parameter Estimation. PE group now applies
+1/(2 weeks) as a threshold. 86400 seconds = 1 day and 14 days = 2 weeks."""